refactor: 쿼리 훅 및 파일 이름 변경
