+++ src/hooks/queries/useQueryUnivOrMajor.ts
